requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
scrapy==2.11.1
crochet==2.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
numpy==1.26.4
openai==1.99.9
tqdm>=4.66
scrapy-playwright>=0.0.32
playwright>=1.45
tqdm>=4.66